+++ subgraph/src/contracts.ts
@@ -6,9 +6,9 @@ import {
 import { Deal } from "../generated/Market/Deal";
 import { Address, Bytes, BigInt } from "@graphprotocol/graph-ts";
 import { ERC20 } from "../generated/Market/ERC20";
-import {Core} from "../generated/Core/Core";
-import {Capacity} from "../generated/Capacity/Capacity";
-import {UNO_BIG_INT, ZERO_BIG_INT} from "./models";
+import { Core } from "../generated/Core/Core";
+import { Capacity } from "../generated/Capacity/Capacity";
+import { MAX_UINT_256, UNO_BIG_INT, ZERO_BIG_INT } from "./models";
 
 // TODO: optimise through multicall contract (currently 2 calls only per token).
 export function getTokenSymbol(address: Bytes): string {
@@ -72,7 +72,7 @@ export function getMinRequiredProofsPerEpoch(contractAddress: Address): BigInt {
   return Capacity.bind(contractAddress).minRequierdProofsPerEpoch();
 }
 
-// It mirrors _failedEpoch in Capacity.sol.
+// It mirrors _failedEpoch internal method in Capacity.sol.
 export function calculateNextFailedCCEpoch(
     maxFailedRatio: BigInt,
     unitCount: BigInt,
@@ -81,6 +81,10 @@ export function calculateNextFailedCCEpoch(
     totalCUFailCount: BigInt,
     lastSnapshotEpoch: BigInt,
   ): BigInt {
+        if (activeUnitCount == ZERO_BIG_INT) {
+            return MAX_UINT_256;
+        }
+
         const maxFails= maxFailedRatio * unitCount;
         let remainingFails = ZERO_BIG_INT;
         if (totalCUFailCount < maxFails) {
@@ -97,8 +101,8 @@ export function calculateNextFailedCCEpoch(
             failedEpoch = UNO_BIG_INT + lastSnapshotEpoch + (remainingFails / activeUnitCount);
         }
 
-        // currently not used.
-        const remainingFailsForLastEpoch = remainingFails % activeUnitCount;
+        // CUrrently it is not used. Artifact from contract side.
+        // const remainingFailsForLastEpoch = remainingFails % activeUnitCount;
         return failedEpoch;
     }